requirements.txt
neo4j>=5.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-adk>=0.1.0
litellm>=1.0.0

